Add a delete endpoint for reports (#192)
* add a delete endpoint for reports

Adds ReportRepository#deleteReport as well as the associated controller
endpoint and typescript client client. Also update the controller
endpoints to all return 500 when an organization is not found rather
than a mix of 404 and 500. Bind UUID directly to the path parameters for
report endpoints.

* Add tests for DatabaseReportRepository#deleteReport

* address PR comments

Don't catch and log in either the repository or controller, play will
handle, don't explicitly return 5XX. Create a manual query for
soft-deleting reports rather than rely on play to avoid an additional
roundtrip to the DB.

* remove unused imports

* only log 'deleted' message when affected rows is nonzero