interfaces for internal JobExecution (#360)
Introduce a new type of repository, JobExecutionRepository, that pulls out all execution related methods from the JobRepository interface.

In the case of Alerts, these two repos will not be backed by the same implementation, and in general that is not a strict requirement and so it makes sense to consider them separately (although in the case of reports, both will still just use the same SQL Ebean impls underneath).

Relatedly, some alert APIs will need to retrieve the state of the most recent execution and so we need an internal model to represent these. We did not have this defined previously because reports are write-only at the moment.

Perhaps somewhat contentiously, I've added another ADT-like hiearchy in JobExecution rather than stuff all fields as top-level Optionals. This prevents you from, for example, attempting to call getResult from an execution that hasn't completed.