remove unused metrics query code and add query format
MetricsController and the associated view model are not currently used
so I'm opting to rip them out just as with the unused alerting code.
Since the time-range won't be used until execution, we also introduce a
serialization format field to the query to allow for dispatch based on
type.

An enum is being used over a sealed hierarchy because no additional
behavior will be shared among MetricsQuery instances.