Add support for periodCount > 1 in PeriodicSchedule (#393)
Add support for PeriodicSchedule instances whose interval is several periods (e.g. 15 minutes instead of just a minute). This only alters the internal model, so for backwards compatibility, the periodCount is set to 1 by default.

I also took this opportunity to refactor the unboundedNextRun code. I removed the check to see if the run was before getRunAtAndAfter because this is already handled in BaseSchedule, and indeed the doc there specifies that subclasses do not need to make this check. This change only a single offset-related test, and I consider that change a bug fix. We were skipping a run that was otherwise valid as it fell between the specified runAtAndAfter and runUntil values.