Create a JobRepository wrapper for alerts (#394)
This wrapper takes an AlertRepository and wraps the returned Alert instances in order to implement the Job interface.

Specifically, a new class AlertExecutionContext is introduced (analogous to ReportExecutionContext) that contains all of the specific dependencies and behavior needed for alert evaluation. AlertJob thus acts as a binding between this context and Alert instances - it's an extremely shallow wrapper on its own.

While we could eventually infer longer execution intervals based on the specifics of the alert query, for now we just poll at a fixed interval (30s) regardless of the alert.