Read-Only Alert Repository (#389)
Introduces a read-only repository for reading alert models that have been serialized to a file. This is essentially just an in-memory repository that hydrates itself from a File - once the repository is opened there is no subsequent File I/O.

UUIDs are computed on the fly when absent to allow for consistency between internal UUID keys and name-based keys. This is fairly straightforward using named-based UUIDs but since JDK8 doesn't include native support for v5 IDs, I've pulled in JUG for this purpose. If this name-based use case becomes more widespread in the org's repos we'll likely pull this up into commons.