Periodically create partitions for alert executions (#398)
Introduce a DailyPartition manager actor that runs within the DatabaseAlertExecutionRepository, periodically attempting to create partitions for the alert executions table that was added in #380. Notably I've introduced another EbeanServer which uses the DDL role (metrics_dba), since that Role is necessary to create tables.

I've included unit tests for the actor while the DB testing is covered by the existing integration tests.