Interface for reloading config-based alert definitions (#430)
FileAlertRepository, introduced in #389, allowed for reading hardcoded alert definitions from a JSON file but had no mechanism for reloading them on updates. While this PR does not actually introduce a working live-update implementation, it does break apart the configuration loading into an interface to make this reloading possible.

The existing static file logic has been extracted into its own ConfigProvider. A PluggableAlertRepsitory configured with a StaticFileConfigProvider would be equivalent in functionality to the previous FileAlertRepository.

Why the additional abstraction if we're not introducing reloading implementations right away? We'll be using this to integrate with Dropbox's live configuration system. I plan on adding a Filesystem watcher implementation later.