Run anti-entropy at JobCoordinator startup (#441)
Run anti-entropy at actor startup

It seems that the JobCoordinator will not perform anti-entropy until
after the first timer fires 1 hour after startup. This ordinarily isn't
a problem because as I understand it the actor will be shuffled
throughout the cluster if the host it resides on is shutdown. In the
cases where an entirely new coordintor is started or all instances are
shut down, the initial wait is unnecessary.