Fix a bug with unbounded schedule fixed points (#447)
Bug-fix for a scheduling issue caught during an end-to-end alert evaluation test.

For times that land exactly on a period, we would return that same period as the next evaluation time. This was because the order of the orElse and Duration::plus operations were inversed, since we always want to perform that plus regardless of if there was a previous run or not. The time returned by the schedule should always be in the future. I've included a test that caught this bug and updated the other tests since their behavior was actually incorrect.