Use raw JDBC conn for DailyPartitionCreator
Ebean's SQLQuery doesn't appear to be working, and SQLUpdate does not
support SELECT statements (which is necessary to call the function).

The existing test suite didn't catch this issue because it created the
partitions for the current date during migrations, and so the partition
creation would silently fail while the tests would still pass. I've
added a new test which creates a scheduled event in the past, which
catches this oversight.