bugfix: add a constraint on the alert id when inserting an execution (#451)
A silly bug, but it's happened nonetheless.

## Expectation
We want to update an alert execution, so we first check it exists by querying for the most recent execution _for this particular ID_.

## Reality
We want to update an alert execution, so we first check it exists by querying for the most recent execution. 

Notice that the above never involves Alert ID.

---

Fortunately, this blows up because we only expect to ever retrieve a single result, instead of several.
The fix is straightforward: add the missing where clause. I've also included a sanity check test that I used to reproduce and identify the issue.