reconstruct JobRef from JobExecutorActor name if possible (#456)
When a `JobExecutorActor` is moved between nodes due to a shard move, it ends up in a sort of purgatory state making no meaningful progress since one or both of the following happen:
1. the state of the actor is not moved, so it doesn't know what job it is responsible for
2. the ticker for the actor isn't restarted, so it doesn't attempt to check if the job needs to run

In order to avoid this entirely, we would need a form of dynamic props to transfer the job ref over when the actor is started post-move.  Unfortunately, classic actor sharding [1] does not allow for this and so we're left with this hack of encoding the ref ID into the actor's name. So now at startup, the actor will attempt to infer its JobRef from its name and start ticking again if it is able to do so. 

[1]: The typed actor implementation of cluster sharding appears to allow for dynamic props, if we ever were to move to this in the future.