decode the actor name before attempting to deserialize (#457)
For whatever reason Akka is not URL encoding the name when run as a
single actor, but it is encoding it in a clustered environment. It
likely does this lazily when the actor name needs to be serialized.

To avoid issues let's just always URL decode before we attempt to use
the name.