separate lookback and period hint, support non-unit periods (#466)
We use period hint for the lookback window on a query, but they should be considered separate concepts. One is the minimum range necessary for data, and the other is the minimum frequency that we should poll this query (these turn out to be the max and min of the aggregator sampling, respectively). We don't actually use the former right now, as all queries are evaluated every minute.

The second change is due to the incorrect assumption made that aggregators will always be a unit value (e.g. exactly 1 hour). There's no reason they shouldn't be, and by using Duration instead of ChronoUnit the lookback should be correctly computed for arbitrary aggregations like 12 hours.

periodHint is renamed to evaluationPeriodHint for clarity.