Batch fetch alerts from postgres (#467)
When we paginate over alerts, we currently fetch their execution status individually. As you can imagine, this is not very performant as we make a round trip to the DB for every alert.

This adds a batch lookup method which fetches the executions in one go. The query itself is much cheaper per-alert (measured up to 1000 records), so including the decrease in network cost I don't expect latency to be a problem again here for a while.