Do not use rollups when evaluating alerts (#469)
We want to avoid the use of rollups in alert evaluation, which could potentially affect correctness. For example, if rollups were continually an hour behind, an hourly sample-aligned query would never see any data.

This adds per-origin (alerts, rest API) controls for enabling rollups. By default rollups are disabled for alerts. If we build sufficient confidence we can re-enable these for alerts in the future.