Add a write-through cache for alert executions (#482)
The cache at the moment is simple, it's just an actor-guarded LRU cache (guava).

An alert execution page lookup will now first attempt to look for all ids in the cache before falling back to the inner
repository (e.g. Postgres). Any cache misses that fell back are written to the cache. Any new successes are written directly to the cache.

This does *not* do any of the following
 - Cache anti-entropy. For the time being the regular evaluation of alerts already serve this function since they happen so frequently.
 - Cache distribution. For a much larger set of alerts, it _could_ make more sense to shard members of the cache and distribute them across the network by using ClusterSharding. It's unclear if this is actually better for most cases since this would require fan-out/fan-in.
 
 I originally attempted to do this using generics to leave this open as a future building block, but backtracked in favor of a specialized implementation. Dealing with generics in Akka messages was not worth it.