make the alert execution cache persistent (#496)
The persistence portion is pretty straightforward:
* Persist all `Put` and `MultiPut` messages.
* Periodically (every 5 minutes) take a snapshot of the entire cache state.

That's it.

The hairy part here was serialization - attempting to serialize the cache directly as a map requires registering a custom mapKey serializer with Jackson (since `CacheKey` isn't a "simple" type). Instead I've created a `CacheSnapshot` type that is essentially a list of `SnapshotEntry`s. The sole purpose of this type is for snapshot ser(de).

Tested locally by running postgres and bringing the server up and down.